Refactored to ports adapter structure
