+++ backend/rag_api/infrastructure/__init__.py
@@ -1,1 +1,1 @@
-# This is an empty __init__.py file to ensure the folder is recognized as a package.
+from . import bindings
