--- backend/rag_api/infrastructure/ports/__init__.py
@@ -1,1 +1,0 @@
-# This is an empty __init__.py file to ensure the folder is recognized as a package.
