--- backend/rag_api/infrastructure/ports/llm_port.py
@@ -1,6 +1,0 @@
-from abc import ABC, abstractmethod
-
-class LLMPort(ABC):
-    @abstractmethod
-    def get_response(self, prompt: str) -> str:
-        pass
