--- backend/rag_api/infrastructure/ports/requirements_store_port.py
@@ -1,10 +1,0 @@
-from abc import ABC, abstractmethod
-
-class RequirementsStorePort(ABC):
-    @abstractmethod
-    def fetch_tickets(self, query: str):
-        pass
-
-    @abstractmethod
-    def convert_to_embeddings(self, tickets):
-        pass
